Issue #401 Support of MariaDB prepared statement metadata caching.
The fix sends metadata caching extended capability flag if server set
it too. If this extended MariaDB capability flag is set, it expects that
there can be no metadata in the Ok packet and uses for query result
the metadata from prepare.
The patch contains the test for metadata consistency.

Since the fix relies on the fix in mysql_common that is not yet merged,
the patch (temporily) changes Cargo.toml to use mysql_common from my
fork.