Added test for the case the packet is longer that max_allowed_packet
i.e. for the case when paramset is split and end in >1 exec bulk
packets.
Also fixed _exec_batch as it would omit the row on which becomes longer
that max allowed packet(i.e. that should go first in the next packet.
Also made "traditional" function to take care of the empty paramset.